fix warning junk
git-svn-id: https://svn.nmap.org/trunk/nmap@441 e0a8ed71-7df4-0310-8962-fdc924857419

+++ targets.c
@@ -421,7 +421,7 @@ gettimeofday(&start, NULL);
 	 if (TIMEVAL_SUBTRACT(t2,time[seq]) > 1000000) {
 	   pt.discardtimesbefore = hostnum;
 	   if (o.debugging) 
-	     printf("Huge send delay: %lu microseconds\n", TIMEVAL_SUBTRACT(t2,t1));
+	     printf("Huge send delay: %lu microseconds\n", (unsigned long) TIMEVAL_SUBTRACT(t2,t1));
 	 }
        }
      } /* for() loop */
